Character Stat을 Gameplay Attribute로 리펙토링
- Enemy와 Player 모두 Stat Component가 아닌 Attribute를 참조하여 스텟 설정

To do
- FP1PlayerStat 구조체는 어떻게 처리?

+++ Source/P1/Character/P1CommonMonster.h
@@ -34,7 +34,7 @@ protected:
 
 
 public:
-	virtual void OnDamaged(int32 Damage, TObjectPtr<AP1Character> Attacker) override;
+	virtual void OnDamaged(float Damage, TObjectPtr<AP1Character> Attacker) override;
 	virtual void OnDead() override;
 	virtual void EnableRagdoll() override;
 	virtual void Disappear() override;
